style(msvc): fix C4244/C4267 in RageDisplay_OGL/Song/ScreenSelectMaster/RageLog/RageFileDriverDeflate/RageDisplay (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226.

- RageDisplay_OGL.cpp (8): a shader-compile helper's GLint/GLsizei
  length/count args fed RString::size(); two int-into-float args to
  LoadMenuPerspective; two std::max<unsigned int> calls fed size_t
  vertex/triangle totals; a size_t old-buffer-size local.
- Song.cpp (7, SS5-protected Song*, characterization-only): four
  reverse-loop size()-1 bounds across RemoveAutoGenNotes/DeleteSteps/
  FreeAllLoadedFromProfile; a lua_createtable/lua_rawseti pair
  serializing background changes to Lua.
- ScreenSelectMaster.cpp (7): a lua_pushnumber(size_t) site; a
  lua_rawgeti fed a size_t loop counter; a SET_POS_PART macro's
  lua_tonumber-into-float (fixed once, covers 3 uses); two
  wrap(int&, size_t) sites; a size_t-into-int local.
- RageLog.cpp (7): a double-into-float timestamp arg; four
  size_t/strlen() values accumulated into unsigned log-buffer
  counters; a std::min<unsigned int> call's second arg; a
  std::min(size_t, size_t) result into an int log-size local.
- RageFileDriverDeflate.cpp (7): a size_t byte-count assigned into
  zlib's uInt avail_out/avail_in fields at 2 sites; two pointer-diffs
  into int (inflate progress tracking); a size_t returned as int from
  WriteInternal.
- RageDisplay.cpp (7): a std::round() result into an int refresh
  rate; a std::ceil() result into an int width; a size_t vertex count
  into StatsAddVerts(int); four mesh-info size_t counts/accumulators
  into int fields.

+++ src/RageDisplay.cpp
@@ -122,7 +122,7 @@ RString RageDisplay::SetVideoMode( VideoModeParams p, bool &bNeedReloadTextures 
 	const DisplayMode supported = d.currentMode() != nullptr ? *d.currentMode() : *d.supportedModes().begin();
 	p.width = supported.width;
 	p.height = supported.height;
-	p.rate = std::round(supported.refreshRate);
+	p.rate = static_cast<int>(std::round(supported.refreshRate));
 	if( (err = this->TryVideoMode(p,bNeedReloadTextures)).empty() )
 		return RString();
 	vs.push_back( err );
@@ -789,7 +789,7 @@ bool RageDisplay::SaveScreenshot( RString sPath, GraphicsFileFormat format )
 		int iHeight = 480;
 		// This used to be lrint. However, lrint causes odd resolutions like
 		// 639x480 (4:3) and 853x480 (16:9). ceil gives correct values. -aj
-		int iWidth = std::ceil( iHeight * GetActualVideoModeParams().fDisplayAspectRatio );
+		int iWidth = static_cast<int>(std::ceil( iHeight * GetActualVideoModeParams().fDisplayAspectRatio ));
 		timer.Touch();
 		RageSurfaceUtils::Zoom( surface, iWidth, iHeight );
 //		LOG->Trace( "%ix%i -> %ix%i (%.3f) in %f seconds", surface->w, surface->h, iWidth, iHeight, GetActualVideoModeParams().fDisplayAspectRatio, timer.GetDeltaTime() );
@@ -893,7 +893,7 @@ void RageDisplay::DrawCompiledGeometry( const RageCompiledGeometry *p, int iMesh
 {
 	this->DrawCompiledGeometryInternal( p, iMeshIndex );
 
-	StatsAddVerts( vMeshes[iMeshIndex].Triangles.size() );
+	StatsAddVerts( static_cast<int>(vMeshes[iMeshIndex].Triangles.size()) );
 }
 
 void RageDisplay::DrawLineStrip( const RageSpriteVertex v[], int iNumVerts, float LineWidth )
@@ -993,10 +993,10 @@ void RageCompiledGeometry::Set( const std::vector<msMesh> &vMeshes, bool bNeedsN
 		MeshInfo& meshInfo = m_vMeshInfo[i];
 		meshInfo.m_bNeedsTextureMatrixScale = false;
 
-		meshInfo.iVertexStart = totalVerts;
-		meshInfo.iVertexCount = Vertices.size();
-		meshInfo.iTriangleStart = totalTriangles;
-		meshInfo.iTriangleCount = Triangles.size();
+		meshInfo.iVertexStart = static_cast<int>(totalVerts);
+		meshInfo.iVertexCount = static_cast<int>(Vertices.size());
+		meshInfo.iTriangleStart = static_cast<int>(totalTriangles);
+		meshInfo.iTriangleCount = static_cast<int>(Triangles.size());
 
 		totalVerts += Vertices.size();
 		totalTriangles += Triangles.size();
